requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.12.0=py36haa95532_0
aiohttp=3.7.3=pypi_0
aiohttp-cors=0.7.0=pypi_0
aioredis=1.3.1=pypi_0
alabaster=0.7.12=py36_0
argh=0.26.2=py36_0
argon2-cffi=20.1.0=py36h2bbff1b_1
astor=0.8.1=py36haa95532_0
astroid=2.5=py36haa95532_1
astunparse=1.6.3=pypi_0
async-timeout=3.0.1=pypi_0
async_generator=1.10=py36h28b3542_0
atomicwrites=1.4.0=py_0
attrs=20.3.0=pyhd3eb1b0_0
autopep8=1.5.6=pyhd3eb1b0_0
babel=2.9.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
bcrypt=3.2.0=py36he774522_0
beautifulsoup4=4.9.3=pypi_0
blas=1.0=mkl
bleach=3.3.0=pyhd3eb1b0_0
blessings=1.7=pypi_0
bokeh=2.2.1=pypi_0
brotlipy=0.7.0=py36h2bbff1b_1003
ca-certificates=2020.12.5=h5b45459_0
cachetools=4.1.1=pypi_0
certifi=2020.12.5=py36ha15d459_1
cffi=1.14.5=py36hcd4344a_0
chardet=4.0.0=py36haa95532_1003
click=7.1.2=pypi_0
cloudpickle=1.6.0=py_0
colorama=0.4.4=pyhd3eb1b0_0
colorful=0.5.4=pypi_0
contextlib2=0.6.0.post1=pypi_0
contextvars=2.4=pypi_0
coverage=5.5=py36h2bbff1b_2
cryptography=3.4.7=py36h71e12ea_0
cudatoolkit=10.2.89=h74a9793_1
cycler=0.10.0=pypi_0
cython=0.29.22=py36hd77b12b_0
dataclasses=0.8=pypi_0
decorator=5.0.5=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
diff-match-patch=20200713=py_0
docutils=0.17=py36haa95532_1
einops=0.3.0=pypi_0
entrypoints=0.3=py36_0
filelock=3.0.12=pypi_0
flake8=3.9.0=pyhd3eb1b0_0
freetype=2.10.4=hd328e21_0
future=0.18.2=py36_1
gast=0.3.3=pypi_0
google=3.0.0=pypi_0
google-api-core=1.23.0=pypi_0
google-auth=1.21.2=pypi_0
google-auth-oauthlib=0.4.1=pypi_0
google-pasta=0.2.0=py_0
googleapis-common-protos=1.52.0=pypi_0
gpustat=0.6.0=pypi_0
grpcio=1.32.0=pypi_0
h5py=2.10.0=py36h5e291fa_0
hdf5=1.10.4=h7ebc959_0
hiredis=1.1.0=pypi_0
humanize=3.0.1=pypi_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=2.10=pyhd3eb1b0_0
idna-ssl=1.1.0=pypi_0
imageio=2.9.0=pypi_0
imagesize=1.2.0=pyhd3eb1b0_0
immutables=0.14=pypi_0
importlib-metadata=3.10.0=py36haa95532_0
importlib_metadata=3.10.0=hd3eb1b0_0
intel-openmp=2020.2=254
intervaltree=3.1.0=py_0
ipykernel=5.3.4=py36h5ca1d4c_0
ipython=7.16.1=py36h5ca1d4c_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.3=pyhd3eb1b0_1
isort=5.8.0=pyhd3eb1b0_0
jedi=0.17.1=py36_0
jinja2=2.11.3=pyhd3eb1b0_0
joblib=0.16.0=pypi_0
jpeg=9b=hb83a4c4_2
jsonpatch=1.26=pypi_0
jsonpointer=2.0=pypi_0
jsonschema=3.2.0=py_2
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_core=4.7.1=py36haa95532_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
keras=2.4.3=pypi_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.2=pyhd3eb1b0_0
keyring=22.3.0=py36haa95532_0
kiwisolver=1.2.0=pypi_0
lazy-object-proxy=1.6.0=py36h2bbff1b_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.14.0=h23ce68f_0
libsodium=1.0.18=h62dcd97_0
libspatialindex=1.9.3=h6c2663c_0
libtiff=4.1.0=h56a325e_1
livelossplot=0.5.3=pypi_0
llvmlite=0.34.0=pypi_0
lz4-c=1.9.3=h2bbff1b_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.3.4=py36haa95532_0
markupsafe=1.1.1=py36he774522_0
matplotlib=3.3.2=pypi_0
mccabe=0.6.1=py36_1
mistune=0.8.4=py36he774522_0
mkl=2020.2=256
mkl-service=2.3.0=py36h196d8e1_0
mkl_fft=1.3.0=py36h46781fe_0
mkl_random=1.1.1=py36h47e9c7a_0
ml-collections=0.1.0=pypi_0
msgpack=1.0.0=pypi_0
msys2-conda-epoch=20160418=1
multidict=5.0.2=pypi_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.0.7=py36_0
nbformat=5.1.3=pyhd3eb1b0_0
nest-asyncio=1.5.1=pyhd3eb1b0_0
networkx=2.5=pypi_0
nibabel=3.1.1=pypi_0
ninja=1.10.2=py36h6d14046_0
notebook=6.3.0=py36haa95532_0
numba=0.51.2=pypi_0
numpy=1.18.5=pypi_0
numpy-base=1.19.2=py36ha3acd2a_0
numpydoc=1.1.0=pyhd3eb1b0_1
nvidia-ml-py3=7.352.0=pypi_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py36_0
opencensus=0.7.11=pypi_0
opencensus-context=0.1.2=pypi_0
opencv-python=4.4.0.42=pypi_0
openssl=1.1.1k=h8ffe710_0
opt-einsum=3.3.0=pypi_0
packaging=20.9=pyhd3eb1b0_0
pandas=1.1.2=pypi_0
pandoc=2.12=haa95532_0
pandocfilters=1.4.3=py36haa95532_1
paramiko=2.7.2=py_0
parso=0.7.0=py_0
pexpect=4.8.0=pyhd3eb1b0_3
pickle5=0.0.11=pypi_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.2.0=py36h4fa10fc_0
pip=21.0.1=py36haa95532_0
pluggy=0.13.1=py36haa95532_0
prometheus_client=0.10.0=pyhd3eb1b0_0
prompt-toolkit=3.0.17=pyh06a4308_0
protobuf=3.13.0=pypi_0
psutil=5.8.0=py36h2bbff1b_1
ptyprocess=0.7.0=pyhd3eb1b0_2
py-spy=0.3.3=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycodestyle=2.6.0=pyhd3eb1b0_0
pycparser=2.20=py_2
pydocstyle=6.0.0=pyhd3eb1b0_0
pyflakes=2.2.0=pyhd3eb1b0_0
pygments=2.8.1=pyhd3eb1b0_0
pylint=2.7.4=py36haa95532_1
pynacl=1.4.0=py36h62dcd97_1
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py36h6538335_2
pyreadline=2.1=py36_1
pyrsistent=0.17.3=py36he774522_0
pysocks=1.7.1=py36haa95532_0
python=3.6.13=h3758d61_0
python-dateutil=2.8.1=pyhd3eb1b0_0
python-deprecated=1.1.0=pypi_0
python-jsonrpc-server=0.3.4=py_1
python-language-server=0.34.1=py36_0
python_abi=3.6=1_cp36m
pytorch=1.6.0=py3.6_cuda102_cudnn7_0
pytz=2021.1=pyhd3eb1b0_0
pywavelets=1.1.1=pypi_0
pywin32=227=py36he774522_1
pywin32-ctypes=0.2.0=py36_1000
pywinpty=0.5.7=py36_0
pyyaml=5.4.1=py36h2bbff1b_1
pyzmq=20.0.0=py36hd77b12b_1
qdarkstyle=3.0.2=pyhd3eb1b0_0
qt=5.9.7=vc14h73c81de_0
qtawesome=1.0.2=pyhd3eb1b0_0
qtconsole=5.0.3=pyhd3eb1b0_0
qtpy=1.9.0=py_0
ray=1.0.1.post1=pypi_0
redis=3.4.1=pypi_0
requests=2.25.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=pypi_0
rope=0.18.0=py_0
rsa=4.6=pypi_0
rtree=0.9.4=py36h21ff451_1
scikit-image=0.17.2=pypi_0
scikit-learn=0.22.2.post1=pypi_0
scipy=1.4.1=pypi_0
seaborn=0.11.0=pypi_0
send2trash=1.5.0=pyhd3eb1b0_1
setuptools=52.0.0=py36haa95532_0
simpleitk=1.2.4=pypi_0
sip=4.19.8=py36h6538335_0
six=1.15.0=py36haa95532_0
snowballstemmer=2.1.0=pyhd3eb1b0_0
sortedcontainers=2.3.0=pyhd3eb1b0_0
soupsieve=2.0.1=pypi_0
sphinx=3.5.3=pyhd3eb1b0_0
sphinxcontrib-applehelp=1.0.2=pyhd3eb1b0_0
sphinxcontrib-devhelp=1.0.2=pyhd3eb1b0_0
sphinxcontrib-htmlhelp=1.0.3=pyhd3eb1b0_0
sphinxcontrib-jsmath=1.0.1=pyhd3eb1b0_0
sphinxcontrib-qthelp=1.0.3=pyhd3eb1b0_0
sphinxcontrib-serializinghtml=1.1.4=pyhd3eb1b0_0
spyder=4.1.5=py36_0
spyder-kernels=1.9.4=py36_0
sqlite=3.35.4=h2bbff1b_0
tabulate=0.8.7=pypi_0
tensorboard=2.3.0=pypi_0
tensorboard-plugin-wit=1.7.0=pypi_0
tensorboardx=2.1=pypi_0
tensorflow=1.14.0=h39a9291_0
tensorflow-base=1.14.0=py36h9f0ad1d_0
tensorflow-estimator=2.3.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.9.4=py36haa95532_0
testpath=0.4.4=pyhd3eb1b0_0
tifffile=2020.9.3=pypi_0
tikzplotlib=0.9.6=pypi_0
tk=8.6.10=he774522_0
toml=0.10.2=pyhd3eb1b0_0
torchfile=0.1.0=pypi_0
torchio=0.17.47=pypi_0
torchnet=0.0.4=pypi_0
torchvision=0.7.0=py36_cu102
tornado=6.1=py36h2bbff1b_0
tqdm=4.49.0=pypi_0
traitlets=4.3.3=py36_0
typed-ast=1.4.2=py36h2bbff1b_1
typing_extensions=3.7.4.3=pyha847dfd_0
ujson=1.35=py36hfa6e2cd_0
umap-learn=0.4.6=pypi_0
urllib3=1.26.4=pyhd3eb1b0_0
vc=14.2=h21ff451_1
visdom=0.1.8.9=pypi_0
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=1.0.2=py36haa95532_1
wcwidth=0.2.5=py_0
webencodings=0.5.1=py36_1
websocket-client=0.57.0=pypi_0
werkzeug=1.0.1=pyhd3eb1b0_0
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=py36_0
win_inet_pton=1.1.0=py36haa95532_0
wincertstore=0.2=py36h7fe50ca_0
winpty=0.4.3=4
wrapt=1.12.1=py36he774522_1
xz=5.2.5=h62dcd97_0
yaml=0.2.5=he774522_0
yapf=0.31.0=pyhd3eb1b0_0
yarl=1.6.3=pypi_0
zeromq=4.3.3=ha925a31_3
zipp=3.4.1=pyhd3eb1b0_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.9=h19a0ad4_0
